Restructure Tab inside Molecule folder & fix import
